requirements.txt
Flask>=2.0.0
python-dotenv>=1.0.0
requests>=2.0.0
pytest>=7.0.0
